Fix another race (in lab0.0/shared_init)
The test harness speaks to QEMU's VMM via FIFO pipes.  It turns out
that when QEMU exist very quickly (which is what happens in the
shared_init case) then this can cause a race with opening these pipes.
I was under the impression that this shouldn't be a problem: the FIFO
should buffer its input and still allow the rest of the system to open
and read from that pipe, but it seems that isn't always the case.

This may have something to do with the fact that QEMU's monitor
emulates a shell over the FIFOs, so it's doing odd things like not
pushing out a full line (ie, there's no newline at the end).

You can still cause this to hang if you do call the poweroff system
call unexpectedly (ie, in the middle of trying to dump VGA memory), so
just don't do that.  Cucumber should handle arbitrary kernel panics
OK, just not arbitrary poweroffs.