Add an hour-based webclock
The idea is that some web applications can have a Cache-Until time
that expires before their content changes.  To simulate this, I've
provided a modified webclock that calculates time every hour, but
still only asks to be cached for a minute.

This will allow you to test an edge-case of your cache: that the cache
correctly handles the cache-until header and the etag header.