prctl thread limit: Now limits the children of the process prctl was called on.
If prctl is called on thread j permitting only x threads per child then j is
unrestricted in the number of threads it can call, but j's children are. Each
child carries its own limit so if j has 2 threads both threads are allowed x
children (actually x-1 as the child of j has a thread that I count as part of
this total. This may be a failure in my implementation.)