Fix ETH194 interrupt dropping bug
I made a typo in some ISR modification code and accidentally set the
whole ISR when I mean to just set a single bit in the ISR.  This
causes the ETH194 to drop any other interrupt types it currently had
queued at the time.

Thanks to James Eady for the fix.