Re-add {R,T}CNT{LO,HI} to eth194
These got dropped during the NE2000 -> ETH194 conversion.  They're
used by some shared code to count transmitted and received packets
(I'm not sure if that's where the "ifconfig" counts come from), so
they should be there.

The behavior is fairly undocumented, but it's just the same as the
NE2000 (where it's also pretty much undocumented).