Remove RTL8029ID{0,1}
These will almost certainly be unused by your driver.  For some reason
they map to the counter registers.  The QEMU code allows you to write
the counter registers, but not to read them.

As there's no mention of this in the ETH194 datasheet, I think it's
probably best to get rid of them...